wip model: Add new wildcard mentioned flags.
This was introduced at feature level 224. Since the server API hasn't
been updated to use "channel" for the enum value, we do the conversion
on the "stream_wildcard_mentioned" flag.

The hard part about it is how we are going to support unreads and later
on render @topic/@channel wildcard mentions correctly. These are
something to work on later on.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>