internal_link [nfc]: Add ApiIsNarrow.
This replaces ApiNarrowIsMentioned and ApiNarrowIsUnread with the
generic ApiNarrowIs element.

This does not change the internal links behavior, hence the nfc.
This is also why this commit hasn't introduced new tests yet.

We convert the IsOperand value into a String when constructing
ApiNarrowIs, so that we don't need to mingle with
ApiNarrowElement.toJson to take care of the enum.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>