compose_box: Use new send button color, shape and icon
DesignVariables.icon's value has been updated to match the current
design. This would affect the appearance of the
ChooseAccountPageOverflowButton on the choose account page, which is
intentional.

Note that we set the color on `Icon` instead of `IconButton`, because
the color shouldn't apply to for example the background color of the
button when highlighted, etc.

See also:

  - https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=3954-13395
  - https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=3862-14350