poll: Make Poll a ChangeNotifier
Previously, we would notify all message list views containing the
message when the poll got updated through a submessage event.

By making Poll a ChangeNotifier, we can limit that notification to only
its listeners to improve performance.  A side-effect of this change is
that we no longer notify listeners if the event is malformed.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>