content: Start rendering poll content
We could also make Poll a subclass of ZulipMessageContent, but that
takes away our ability to seal ZulipMessageContent.  Making a wrapper
class around Poll is easy enough for keeping that benefit.

_reparseContent currently happens when there is content edit, which is
forbidden for poll messages, (and you can't turn an existing message
into a poll).  In most cases the same `PollContent` instance will be
used for each poll message, unlike `ZulipContent`, and a poll message
never needs to get reparsed.

Fixes: #165

Signed-off-by: Zixuan James Li <zixuan@zulip.com>