compose_box: Implement input redesign
The bottom underline implementation draws similarity to that of
`Spoiler`'s.  We could have used `UnderlineInputBorder` on
`InputDecoration`, but that also comes with other input state styles
that we do not need (e.g.: focused, disabled, etc.).

Note that we use `withFadedAlpha` on `designVariables.textInput` because
the color is already transparent in dark mode, and the helper allows us
to multiply, instead of to override, the alpha channel of the color with
a factor.

`ClipRect`'s size is determined by the `ConstrainedBox`.  This is
mainly for showing the content through the `contentPadding` of the
`TextField`, so that our `InsetShadowBox` can fade it smoothly there.
The shadow is always there, but it is only visible when the
`TextField` is long enough to be scrollable.

See also:

  - https://github.com/zulip/zulip-flutter/pull/928#discussion_r1815661095
  - https://github.com/zulip/zulip-flutter/pull/928#issuecomment-2335042432,
    which elaborates on the issue we intend to solve with the
    `ClipRect` setup.
  - https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=3954-13395
  - https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=3862-14350

Fixes: #915

Signed-off-by: Zixuan James Li <zixuan@zulip.com>