poll: Support vote/unvote for polls
Visually, this does not change the appearence of the vote count box.

This does not implement local echoing for voting. Instead, we rely on
the submessage events to get the updates after voting.

For accessbility, the touch target is larger than the button.
See also:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/Poll.20vote.2Funvote.20UI/near/1952724

We add a TODO for pending poll votes visual indicator.  This would also apply
to emoji reaction pills.
See also:
  https://github.com/zulip/zulip-flutter/pull/939#pullrequestreview-2355412278

Fixes: #166

Signed-off-by: Zixuan James Li <zixuan@zulip.com>