poll: Enlarge and vetically center vote count box
Previously, there were only bottom and end paddings; we redistribute
the vertical paddings so the vote count box is padded on both sides.

The end padding remains 5px because we want to keep the options aligned
to the start with the surrounding messages.

This change also affects the poll question, where 2.5px of space gets
moved to the first option row.

Similarly, the end of the poll was previously padded by the last poll
option alone. Now 2.5px of space gets moved to the `Padding` wrapping
the entire `Column` of the poll.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>