compose_box: Notify interactions with compose buttons
The new _ChannelComposeButtonRow needs to be stateful because changes to
the topic input affects the typing destination.  This is similar to
_StreamContentInput.

We send "typing started" notices as long as the user presses any of the
compose buttons, regardless if they proceed to attach files or not.
This will later apply to other kinds of compose buttons, too.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>