compose: Disable compose box widgets when sending message
Also clear the content input only after the message is successfully
sent.

For the send button, unlike when there are validation errors, having
an error dialog when its disabled for sending a message feels a bit
redundant and doesn't help much.  This can be annoying when the user
misclicks and hit the send button multiple times.

This implements some behavior not specified in Figma. When disabled:

- Because the height of the progress indicator is 2px, to avoid shifting
  the content of the compose box, we expand it by 2px in height instead
  of 1px.

- The bar containing only buttons other than the send button is faded,
  because the send button can be disabled separately through validation
  errors.  If we need to disable the other compose buttons individually in the
  future, we can control the styling of the buttons instead.

The progress indicator will be added in a later commit.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>