compose [nfc]: Move "there is none" logic out of ComposeBox widget
It's a little confusing to see a widget named ComposeBox returning
something that's neither a compose box nor a substitute/placeholder
for one, but something (SizedBox.shrink) that's meant to show
nothing at all. So, pull that logic out to the caller, which nicely
puts it near some related logic (the `hasComposeBox` condition for a
MediaQuery.removePadding).