compose [nfc]: Make state "has-a" instead of "is-a" ComposeBoxController
We've been using ComposeBoxController as an interface, implemented
by _StreamComposeBoxState and _FixedDestinationComposeBoxState.

This commit puts the relevant implementation in ComposeBoxController
itself -- really in subclasses StreamComposeBoxController and
FixedDestinationComposeBoxController -- and makes those _FooState
classes instantiate it and store the instance in a field.

Next, we'll move the field up from those _FooState classes onto
ComposeBox's state to make the controller straightforwardly
available there (i.e. available without going through GlobalKey
logic). That will be helpful for #720 when the controller grows a
send-in-progress flag; ComposeBox will be a conveniently central
place to build a progress indicator for that state.