compose [nfc]: Pass whole `controller` down (1/6); _StreamContentInput
For #720, we'd like to add fields to ComposeBoxController and use
them in a bunch of widgets deeper in the compose box, like the
various buttons, without having to add references to them in those
widgets' constructors. With that in mind, pass the whole controller
down, and the widgets can access the fields directly from it. As
suggested by Greg:
  https://github.com/zulip/zulip-flutter/pull/1090#discussion_r1860183031