msglist: Show topic visibility on app bar
The referenced design from the legacy mobile app displays
the 'mute' icon when the channel is muted, replacing a privacy
level related icon (e.g.: web public, invite only):
  https://github.com/zulip/zulip-mobile/blob/a115df1f71c9dc31e9b41060a8d57b51c017d786/src/streams/StreamIcon.js#L20-L29
We diverge from that in the Flutter app.  We will have a different
place to show channel mute/unmute status in #347.

This implementation also shows the corresponding icons for 'muted' and
'unmuted' topics; previously, only the icon for 'follow' was displayed.
And we continue using the existing icons in the Flutter app, without
trying to match the exact ones in the old design.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>