db: Start generating schema versions for migrations
Generating schema_versions.g.dart is crucial because we
otherwise only have access to the latest schema when running migrations.

That would mean that if we later drop or alter a table or column
mentioned in an existing migration, the meaning of the existing
migration would change. The affected existing migration would then
migrate to an unintended state that doesn't match what later
migrations expect, and it or a later migration might fail.

See also discussion:
  https://github.com/zulip/zulip-flutter/pull/1248#discussion_r1960988853

---

An alternative to using all these commands for generating files is
`dart run drift_dev make-migrations`, which is essentially a wrapper
for the `schema {dump,generate,steps}` subcommands.  `make-migrations`
let us manage multiple database schemas by configuring them with
`build.yaml`, and it dictates the which subdirectories the generated
files will be created at.

Because `make-migrations` does not offer the same level of
customizations to designate exactly where the output files will be,
opting out from it for now.

We can revisit this if it starts to offer features that are not
available with the subcommands, or that we find the need for managing
multiple databases.

See also:
  https://drift.simonbinder.eu/migrations/step_by_step/#manual-generation

Signed-off-by: Zixuan James Li <zixuan@zulip.com>