api [nfc]: Introduce TopicName.displayName, and use where needed
Each of these call sites will need to be updated for #1250.  Those
sites can now be found by listing the references to this getter.

As a bonus the type-checker will point out most of the places that
need updates, because this getter will become nullable.  (The
exceptions are where it's used in a string interpolation, because
`null.toString()` is valid and returns "null".)