api: Parse UpdateMessageMoveData with inter-related constraints
This moves some assertions so that we can make all fields
non-nullable, with reasonable fallback values.  As of writing, we do not
use origStreamId (a.k.a.: 'stream_id') when there was no message move,
even though it is present if there were content edits.  This makes
dropping 'stream_id' when parsing `moveData` into `null` acceptable for
now.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>