api: Extract and parse UpdateMessageMoveData from UpdateMessageEvent
This is mostly NFC except that we were logging and ignoring malformed
event data before, and now start to throw errors within the parsing
code.

This data structure encapsulates some checks so that we can make all
fields non-nullable, with reasonable fallback values.  As of writing,
we do not use origStreamId (a.k.a.: 'stream_id') when there was no
message move, even though it is present if there were content edits
This makes dropping 'stream_id' when parsing `moveData` into `null`
acceptable for now.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>