unreads: Handle updates when there are moved messages
We could have extend `_removeAllInStreamTopic` to fit the needs of
handling moved unreads, but it is used in a code path where messages
are marked as read (which will become hotter after #81).

The new helper is designed to avoid making copies of message IDs
when possible.  As for the name, we use "pop" (inspired by Python)
to indicate that the removed items are returned, since "removeAll"
in Dart doesn't carry that meaning.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>