compose: Support save snippets
We sort the saved snippets by title to be consistent with the web
implementation.

A subtle nuance of having a compose box in a modal bottom sheet is that
the message list page behind will shift when the keyboard is expanded,
even though its compose box is not visible anyway.

An UX improvement would be preserving the inputs on the saved snippet
compose box after the user has navigated away.

Fixes: #863