app: Translate app title
The title right now is just "Zulip", which on its own should translate to
"Zulip" in other languages. But because this is user-visible, we make it
translatable, as we can always change it to a phrase or anything else.

This string is applicable for Android, not iOS. The description,
adapted from the dartdoc of [WidgetsApp.title], for translator mentions
this fact to describe context in which the string appears.

While this might also appear on other platforms like web, there is no
need to bring this up as we do not currently support them.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>