compose: Translate message for a filename with size
The string is used at the end of the "errorFilesTooLarge" message,
which includes a list of files with its size that are too large.

We didn't mention that it is a part of the error message because this
part of the context is not as relevant as that it fits in a single
line.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>