content [nfc]: Follow browser preference setting when opening links
This preserves the previous behavior (external application on iOS,
and in-app browser on Android) when browserPreference is `null`.

There is no way for the client to update the setting for now, hence the
NFC.

The helper could have been a static method instead of an extension,
similar to ThemeSetting.displayName, but that will be a lot more verbose
for the caller.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>