compose: Only show "general chat" after picking the topic
Previously, "Message #stream > general chat" was the hint text for
content input as long as the topic input is empty and topics are not
mandatory.

Showing the default topic does not help create incentive for the user
to pick a topic first.  So only show it when they intend to leave the
topic empty.

See discussion:
  https://chat.zulip.org/#narrow/channel/530-mobile-design/topic/general.20chat.20design.20.23F1297/near/2088870

Signed-off-by: Zixuan James Li <zixuan@zulip.com>