message: Identify moves by inequality new vs orig, not just non-nullity
This is NFC except in certain malformed cases: where either
newStreamId is present and equals origStreamId, or newTopic is
present and equals origTopic.  The API docs say neither of those
should happen.

This version is useful because it brings this code closer to how it
will look when we move this logic of "were the messages moved, and
how" into the API parsing code so that other parts of the data model
can reuse the results.