message [nfc]: In moves, give newStreamId/newTopic their fallback values sooner
To see that this is NFC, first observe that the only possible
behavior change this could cause is to change some event from being
ignored (hitting a `return` in this section of code) to being
processed by the code below, or vice versa.

(It might also change what particular message gets printed to the
debug log; but that doesn't count as a behavior change.)

Then both before and after this change, the events that this overall
set of conditions accepts are those that (a) have all of origTopic,
origStreamId, and propagateMode non-null, and (b) have at least one
of newTopic or newStreamId with a non-null value different from
origTopic or origStreamId respectively.

This change is useful because it brings this logic closer to how it
will look when we move it into the API parsing code.

While here, we add a debug log message on unexpected propagateMode,
to further bring this closer to how it will look in the API parsing
code.

Co-authored-by: Zixuan James Li <zixuan@zulip.com>