topics: Add topic list page
The topic-list page app bar should match the Figma design except for its
height.  That part needs to be coordinated with other pages.

Also on the topic-list page app bar, the icon "chveron_down.svg"
(originally "chveron-down-16.svg") comes from:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=6820-36184&m=dev

and not the "Icons" page, since the "chveron-down" there
appears to be different from the one on the topic list page design.

We also leave out the new topic button and topic filtering.  Those are
out-of-scope for #1158.

---

On the message-list page, "TOPICS" is not aligned to the middle part
of the app bar, that's because we haven't got to #1039 yet.

---

The topic-list implementation is quite similar to parts of inbox page
and message-list page.  Therefore, we structure the code to make them
look similar to compare for changes side-by-side to help with reviewing
what has changed.

Figma design:
  https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=6819-35869&m=dev