inbox: Prepare header/DM/topic items for topic-list redesign
In the new design, topic items on the inbox page have a lot in common
with topic items on the topic-list page.  This applies some of the
changes in the new inbox page design, to prepare for reusing code on
the topic-list page later.

---

What has changed to match the inbox page design, and how they relate to
the topic-list design:

- The topic visibility and at-mention icon markers on topic/DM items.
  These are identical the topic-list design.

- The arrangement of text and icons on the topic items, and handling of
  resolved topics.  These are identical to the topic-list design.

- The at-mention icon markers on header items and the arrangement of
  text and trailing icons on DM items.
  These do not apply to the topic-list, but we changed them anyway so
  that the inbox page looks consistent.

- The removal of the Mateiral Design splash effect.

What has not changed to match the new inbox page design (since inbox
page redesign is not the primary focus):

- The DM item icons (i.e. user avatar for 1-1 DM, and group DM icons).

- The arrangement of leading icons and text on the header items, and
  their background color.

- The unread markers.  For these, the topic-list page design is
  different.

- Gestures on header/DM/topic inbox items.