db: Store language global setting in database
An alternative to implementing _fromLanguageTag is using the
LocaleParser from package:intl/locale.dart.  In this case though,
the values to parse always come from `toLanguageTag`, so we can
utilize that knowledge to get away with a more simple implementation,
that should also be handy upstream.

The locale stored in the database is not guaranteed to be one of our
supported value, especially if we need to rename locales in the future.
See discussion:
  https://github.com/zulip/zulip-flutter/pull/1513#issuecomment-2899633541